DOC #14 apply autoapi
