--- docs/source/api.rst
@@ -1,24 +1,0 @@
-.. _api:
-
-===
-API
-===
-
-Application-Programmer Interface - the source code.
-
-.. autosummary::
-   :toctree: generated
-
-   ~hklpy2.backends
-   ~hklpy2.blocks
-   ~hklpy2.diffract
-   ~hklpy2.geom
-   ~hklpy2.misc
-   ~hklpy2.ops
-   ~hklpy2.user
-   ~hklpy2.wavelength_support
-
-.. figure:: _static/hklpy2-block-diagram.png
-   :alt: Block hklpy2-block-diagram
-
-   Functional blocks in  |hklpy2|.
