+++ hklpy2/blocks/__init__.py
@@ -2,7 +2,6 @@
 Blocks support Diffractometer Operations.
 
 .. autosummary::
-   :toctree: generated
 
    ~configure
    ~constraints
