+++ hklpy2/blocks/sample.py
@@ -2,7 +2,6 @@
 A Crystalline Sample.
 
 .. autosummary::
-    :toctree: generated
 
     ~Sample
     ~SampleError
