+++ hklpy2/blocks/tests/test_constraints.py
@@ -2,7 +2,7 @@ from contextlib import nullcontext as does_not_raise
 
 import pytest
 
-from ...geom import creator
+from ...diffract import creator
 from ...misc import ConfigurationError
 from ...tests.common import assert_context_result
 from ..constraints import ConstraintBase
