+++ hklpy2/blocks/tests/test_reflection.py
@@ -2,7 +2,7 @@ from contextlib import nullcontext as does_not_raise
 
 import pytest
 
-from ...geom import creator
+from ...diffract import creator
 from ...misc import ConfigurationError
 from ...tests.common import assert_context_result
 from ...tests.models import add_oriented_vibranium_to_e4cv
