+++ hklpy2/tests/test_backends.py
@@ -1,9 +1,9 @@
 import pytest
 
 from .. import SolverBase as AppBase
-from .. import get_solver
 from ..backends.base import SolverBase as BackendBase
 from ..backends.no_op import NoOpSolver as BackendSolver
+from ..misc import get_solver
 from . import NO_OP_SOLVER_TYPE_STR
