+++ hklpy2/tests/test_user.py
@@ -6,7 +6,7 @@ import numpy.testing
 import pytest
 
 from ..blocks.lattice import SI_LATTICE_PARAMETER
-from ..geom import creator
+from ..diffract import creator
 from ..misc import ReflectionError
 from ..ops import OperationsError
 from ..user import add_sample
