+++ hklpy2/wavelength_support.py
@@ -4,7 +4,6 @@ Wavelength of the incident radiation.
 Supports compatible unit conversions.
 
 .. autosummary::
-   :toctree: generated
 
     ~A_KEV
     ~ConstantMonochromaticWavelength
